improvements to enemy movement made
Stopping distance turned to 0, currently this causes enemy to try be in exact location as player but when attack mode is added, this will be fixed. Stopping distance was causing issues of the enemy never being within 1 unit of a location. Enemy now sets desPointSet to false within 3 units (just in case a location is actually inaccessible then the enemy will look for a new one when fairly close). Rigid body taken off of the enemy prefab to prevent issues with navmesh and physics interacting.