added enemy attack
Enemy can now attack when player is in range. Enemy colour changes black when in attack mode and yellow when the attack is actively happening. If the enemy hits a player, "hitting" is printed to the console. Rigid body had to be added back to the enemy to allow the collider to detect hits correctly, but it has been made kinematic and ignores gravity to prevent issues of physics also affecting the enemies movement.